pyproject.toml
[project]
name = "knowledge-os"
version = "0.1.0"
description = "Knowledge OS Core - Self-hosted, local-first, agent-aware, policy-driven"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "psycopg>=3.2.0",
    "psycopg-pool>=3.2.0",

    # Pydantic v2
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",

    # Observability
    "structlog>=24.0.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-exporter-otlp>=1.25.0",

    # CORS
    "python-multipart>=0.0.9",

    # YAML config
    "pyyaml>=6.0.0",

    # Git integration
    "python-git>=1.0.0",

    # Agent runtime
    "pydanticai>=0.0.0",

    # Retrieval
    "pgvector>=0.3.0",
    "httpx>=0.28.0",
    "tiktoken>=0.7.0",

    # Research Runtime
    "crawl4ai>=0.4",
    "docling>=2.0",
    "playwright>=1.50",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501", "B008"]  # B008: Query/Depends in FastAPI routes is standard pattern, not a bug

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
